pyproject.toml
[tool.poetry]
name = "wordle"
version = "0.1.0"
description = ""
authors = ["jplamb"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
beautifulsoup4 = "^4.12.2"
nltk = "^3.8.1"
pandas = "^2.0.1"
openpyxl = "^3.1.2"
mod-wsgi = "^4.9.4"
mysqlclient = "^2.1.1"
django = "^4.2.1"
djangorestframework = "^3.14.0"
django-webpack-loader = "^1.8.1"
django-cors-headers = "^3.14.0"
django-rest-auth = "^0.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
